Module partiql_source_map::location

Types representing positions, spans, locations, etc of parsed PartiQL text.

Structs

ByteOffset

A 0-indexed byte offset, relative to some other position.

BytePosition

A 0-indexed byte absolute position (i.e., relative to the start of a &str)

CharOffset

A 0-indexed char offset, relative to some other position.

LineAndCharPosition

A 0-indexed line & char absolute position (i.e., relative to the start of a &str)

LineAndColumn

A 1-indexed line and column location intended for usage in errors/warnings/lints/etc.

LineOffset

A 0-indexed line offset, relative to some other position.

Located

A wrapper type that holds an inner value and a location for it

Location

A range with an inclusive start and exclusive end.

Traits

ToLocated

Trait adding a to_located method to ease construction of Located from its inner value.